--- js/form-toggle.js
@@ -1,21 +1,0 @@
-const switchToInactiveState = (form) => {
-  form.classList.add(`${form.classList[0]}--disabled`);
-
-  const childrenList = form.childNodes;
-
-  for (let i = 0; i < childrenList.length; i++) {
-    childrenList[i].disabled = true;
-  }
-};
-
-const switchToActiveState = (form) => {
-  form.classList.remove(`${form.classList[0]}--disabled`);
-
-  const childrenList = form.childNodes;
-
-  for (let i = 0; i < childrenList.length; i++) {
-    childrenList[i].disabled = false;
-  }
-};
-
-export {switchToInactiveState, switchToActiveState};
