+++ js/main.js
@@ -1,8 +1,7 @@
 import './ad.js';
 import './similar-ads.js';
-import './form-toggle.js';
-import {switchToInactiveState, switchToActiveState} from './form-toggle.js';
 import './form.js';
+import {switchToInactiveState, switchToActiveState} from './form.js';
 
 const adForm = document.querySelector('.ad-form');
 switchToInactiveState(adForm);
